\name{X}
\alias{X}
\docType{data}
\title{
  An example data set.
}
\description{
  An example
}
\usage{data(X)}
\format{
  The format is:
 num [1:3, 1:5] 1 2 0 2 2 1 1 2 2 2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:5] "A" "B" "C" "D" ...
}
\examples{
data(X)
## maybe str(X) ; plot(X) ...
}
\keyword{datasets}
