\name{tht}
\alias{tht}
\title{
  Andrews path function
}
\description{
  Make an Andrews path function
}
\usage{
tht(p, t)
}
\arguments{
  \item{p}{
    The dimension.
  }
  \item{t}{
    The time point.
  }
}
\value{
  p-vector of Andrews path function at time t.
}
\references{
  Kwak and Huh(2008) Andrews Plots for Extended Uses, The Korean Communications in Statistics.
}
\author{
  Il-Youp Kwak
}
